requirements.txt
streamlit==1.32.0
google-search-results==2.4.2
python-dotenv==1.0.1
agno==0.1.0 